Chaos! manage_species_lists was more like an observation action.
Moved to new Observations::SpeciesListsController
