+++ app/controllers/species_lists/observations_controller.rb
@@ -1,67 +1,103 @@
 # frozen_string_literal: true
 
-#   :manage_species_lists
-# Linked from observation_tabs_helper, maybe should also link from spl :show?
-# Table of links for dealing with a list of obs line by line, can add or remove
 #
+#   :add_remove_observations
+# Linked from observations :index and species_lists :show
+# Form that sends all obs from current query to be added or removed from an spl
+
 module SpeciesLists
   class ObservationsController < ApplicationController
     before_action :login_required
 
-    ###########################################################################
-    # :manage_species_lists
-    #
-    # Form (table of post_button links) to let user add/remove one observation
-    # at a time from a species list. Maybe use param[:commit] like above?
+    # :add_remove_observations
+    # Form to add or remove the current *query* of observations
     def edit
-      @observation = find_or_goto_index(Observation, params[:id].to_s)
-      @all_lists = @user.all_editable_species_lists
+      pass_query_params
+      @id = params[:species_list].to_s
+      @query = find_obs_query_or_redirect
     end
 
-    # new endpoint for :add_observation_to_species_list and
-    # :remove_observation_from_species_list. use params[:commit]
+    # :post_add_remove_observations
+    # PUT endpoint to either add or remove
+    # a *query* of observations from a species list
     def update
-      return unless (species_list = find_species_list!)
+      pass_query_params
+      id = params[:species_list].to_s
+      spl = find_list_or_reload_form(id)
+      return unless spl
 
-      return unless (observation = find_observation!)
+      query = find_obs_query_or_redirect(spl)
+      return unless query
 
-      unless check_permission!(species_list)
-        return redirect_to(species_list_path(species_list.id))
+      do_add_remove_observations_by_query(spl, query)
+      redirect_to(species_list_path(spl.id))
+    end
+
+    private
+
+    def find_obs_query_or_redirect(spl = nil)
+      query = find_query(:Observation)
+      return query if query
+
+      flash_error(:species_list_add_remove_no_query.t)
+      if spl
+        redirect_to(species_list_path(spl.id))
+      else
+        redirect_to(species_lists_path)
       end
+      nil
+    end
+
+    def find_list_or_reload_form(id)
+      list = lookup_species_list_by_id_or_name(id)
+      return list if list
+
+      flash_error(:species_list_add_remove_bad_name.t(name: id.inspect))
+      redirect_to(add_query_param(action: :edit, species_list: id))
+      nil
+    end
+
+    def lookup_species_list_by_id_or_name(str)
+      if /^\d+$/.match?(str)
+        SpeciesList.safe_find(str)
+      else
+        SpeciesList.find_by(title: str)
+      end
+    end
+
+    def do_add_remove_observations_by_query(spl, query)
+      return unless check_permission!(spl)
 
       if params[:commit] == :ADD.l
-        add_observation_to_species_list(species_list, observation)
+        do_add_observations_by_query(spl, query)
       elsif params[:commit] == :REMOVE.l
-        remove_observation_from_species_list(species_list, observation)
+        do_remove_observations_by_query(spl, query)
       else
         flash_error("Invalid mode: #{params[:commit].inspect}")
       end
     end
 
-    private
+    def do_add_observations_by_query(species_list, query)
+      ids = query.result_ids - species_list.observation_ids
+      return if ids.empty?
 
-    # Used by manage_species_lists.
-    def remove_observation_from_species_list(species_list, observation)
-      species_list.remove_observation(observation)
-      flash_notice(:runtime_species_list_remove_observation_success.
-        t(name: species_list.unique_format_name, id: observation.id))
-      redirect_to(action: :edit, id: observation.id)
+      # This is apparently extremely inefficient.  Danny says it times out for
+      # large species_lists, such as "Neotropical Fungi".
+      # species_list.observation_ids += ids
+      SpeciesListObservation.insert_all(
+        ids.map do |id|
+          { observation_id: id, species_list_id: species_list.id }
+        end
+      )
+      flash_notice(:species_list_add_remove_add_success.t(num: ids.length))
     end
 
-    # Used by manage_species_lists.
-    def add_observation_to_species_list(species_list, observation)
-      species_list.add_observation(observation)
-      flash_notice(:runtime_species_list_add_observation_success.
-        t(name: species_list.unique_format_name, id: observation.id))
-      redirect_to(action: :edit, id: observation.id)
-    end
+    def do_remove_observations_by_query(species_list, query)
+      ids = query.result_ids & species_list.observation_ids
+      return if ids.empty?
 
-    def find_species_list!
-      find_or_goto_index(SpeciesList, params[:species_list])
-    end
-
-    def find_observation!
-      find_or_goto_index(Observation, params[:observation])
+      species_list.observation_ids -= ids
+      flash_notice(:species_list_add_remove_remove_success.t(num: ids.length))
     end
   end
 end
