--- app/views/species_lists/observations/manage_species_lists.html.erb
@@ -1,79 +1,0 @@
-<%
-  @title = :species_list_manage_title.t(name: @observation.unique_format_name)
-
-  tabs = [
-    link_to(:cancel_and_show.t(type: :observation), @observation.show_link_args)
-  ]
-  @tabsets = { right: draw_tab_set(tabs) }
-
-  obs_lists = []
-  other_lists = []
-  @all_lists.each do |list|
-    if list.observations.member?(@observation)
-      obs_lists << list
-    else
-      other_lists << list
-    end
-  end
-%>
-
-<% if obs_lists.any? %>
-  <p><%= :species_list_manage_belongs_to.t %>:</p>
-  <table class="table-striped table-manage-species-lists">
-    <% obs_lists.each do |species_list| %>
-      <tr>
-
-        <td>
-          <%=
-            when_span = content_tag(:span, species_list.when, class: :ListWhen)
-            what_span = content_tag(:span, species_list.unique_text_name, class: :ListWhat)
-            link_to(when_span + ": " + what_span, species_list.show_link_args)
-          %><br/>
-          <span class="ListWhere"><%= species_list.place_name.t %></span> |
-          <span class="ListWho"><%= user_link(species_list.user) %></span>
-        </td>
-
-        <td>
-          <%= if check_permission(species_list)
-            link_to(:REMOVE.t,
-                    { action: :remove_observation_from_species_list,
-                      observation: @observation.id,
-                      species_list: species_list.id},
-                    data: { confirm: :are_you_sure.l })
-          end %>
-        </td>
-
-      </tr>
-    <% end %>
-  </table>
-<% end %>
-
-<% if other_lists.any? %>
-  <p class="mt-3"><%= :species_list_manage_doesnt_contain.t %>:</p>
-  <table class="table-striped table-manage-species-lists">
-    <% other_lists.each do |species_list| %>
-      <tr>
-
-        <td>
-          <%=
-            when_span = content_tag(:span, species_list.when, class: :ListWhen)
-            what_span = content_tag(:span, species_list.unique_text_name, class: :ListWhat)
-            link_to(when_span + ": " + what_span, species_list.show_link_args)
-          %><br/>
-          <span class="ListWhere"><%= species_list.place_name.t rescue :UNKNOWN.t %></span> |
-          <span class="ListWho"><%= user_link(species_list.user) %></span>
-        </td>
-
-        <td>
-          <%= if check_permission(species_list)
-            link_to(:ADD.t,
-                    { action: :add_observation_to_species_list,
-                      observation: @observation.id,
-                      species_list: species_list.id})
-          end %>
-        </td>
-
-      </tr>
-    <% end %>
-  </table>
-<% end %>
