+++ test/controllers/species_lists/observations_controller_test.rb
@@ -4,103 +4,134 @@ require("test_helper")
 
 module SpeciesLists
   class ObservationsControllerTest < FunctionalTestCase
-    def test_manage_species_lists
-      obs = observations(:coprinus_comatus_obs)
-      params = { id: obs.id.to_s }
-      requires_login(:manage_species_lists, params: params)
+    def test_add_remove_observations
+      query = Query.lookup(:Observation, :all, users: users(:mary))
+      assert(query.num_results > 1)
+      params = @controller.query_params(query)
 
-      assert(assigns_exist, "Missing species lists!")
+      requires_login(:edit)
+      assert_response(:redirect)
+      assert_redirected_to(species_lists_path)
+      assert_flash_error
+
+      get(:edit, params: params)
+      assert_response(:success)
+      assert_input_value(:species_list, "")
+
+      # I wonder if this will work
+      get(:edit, params: params.merge(species_list: "blah"))
+      assert_response(:success)
+      assert_input_value(:species_list, "blah")
     end
 
-    def test_add_observation_to_species_list
-      sp = species_lists(:first_species_list)
-      obs = observations(:coprinus_comatus_obs)
-      assert_not(sp.observations.member?(obs))
-      params = { species_list: sp.id, observation: obs.id }
-      requires_login(:add_observation_to_species_list, params)
-      assert_redirected_to(action: :manage_species_lists, id: obs.id)
-      assert(sp.reload.observations.member?(obs))
-    end
+    def test_post_add_remove_observations
+      query = Query.lookup(:Observation, :all, users: users(:mary))
+      assert(query.num_results > 1)
+      params = @controller.query_params(query)
 
-    def test_add_observation_to_species_list_no_permission
-      sp = species_lists(:first_species_list)
-      obs = observations(:coprinus_comatus_obs)
-      assert_not(sp.observations.member?(obs))
-      params = { species_list: sp.id, observation: obs.id }
-      login("dick")
-      get(:add_observation_to_species_list, params: params)
-      assert_redirected_to(species_list_path(sp.id))
-      assert_not(sp.reload.observations.member?(obs))
-    end
-
-    def test_remove_observation_from_species_list
       spl = species_lists(:unknown_species_list)
-      obs = observations(:minimal_unknown_obs)
-      assert(spl.observations.member?(obs))
-      params = { species_list: spl.id, observation: obs.id }
-      owner = spl.user.login
-      assert_not_equal("rolf", owner)
+      old_count = spl.observations.size
+      new_count = (spl.observations + query.results).uniq.count
 
-      # Try with non-owner (can't use requires_user since failure is a redirect)
-      # effectively fails and gets redirected to show_species_list
-      requires_login(:remove_observation_from_species_list, params)
+      # make sure there are already some observations in list
+      assert(old_count > 1)
+      # make sure we are actually trying to add some observations!
+      assert(new_count > old_count)
+      # make sure some of the query results are already in there
+      assert(query.results & spl.observations != [])
+
+      put_requires_login(:update)
+      assert_response(:redirect)
+      assert_redirected_to(
+        edit_species_list_observations_path(species_list: spl.id)
+      )
+      assert_flash_error
+      assert_equal(old_count, spl.reload.observations.size)
+
+      put(:update, params: params)
+      assert_response(:redirect)
+      assert_redirected_to(
+        edit_species_list_observations_path(species_list: spl.id)
+      )
+      assert_flash_error
+      assert_equal(old_count, spl.reload.observations.size)
+
+      put(:update, params: params.merge(species_list: "blah"))
+      assert_response(:redirect)
+      assert_redirected_to(
+        edit_species_list_observations_path(species_list: spl.id)
+      )
+      assert_flash_error
+      assert_equal(old_count, spl.reload.observations.size)
+
+      put(:update, params: { species_list: spl.title })
+      assert_response(:redirect)
       assert_redirected_to(species_list_path(spl.id))
-      assert(spl.reload.observations.member?(obs))
+      assert_flash_error
+      assert_equal(old_count, spl.reload.observations.size)
 
-      login(owner)
-      get(:remove_observation_from_species_list, params: params)
-      assert_redirected_to(action: "manage_species_lists", id: obs.id)
-      assert_not(spl.reload.observations.member?(obs))
+      put(:update, params: params.merge(species_list: spl.title))
+      assert_response(:redirect)
+      assert_redirected_to(species_list_path(spl.id))
+      assert_flash_error
+      assert_equal(old_count, spl.reload.observations.size)
+
+      put(:update,
+          params: params.merge(commit: "bogus", species_list: spl.title))
+      assert_response(:redirect)
+      assert_redirected_to(%r{/species_list/show_species_list})
+      assert_flash_error
+      assert_equal(old_count, spl.reload.observations.size)
+
+      put(:update,
+          params: params.merge(commit: :ADD.l, species_list: spl.title))
+      assert_response(:redirect)
+      assert_redirected_to(%r{/species_list/show_species_list})
+      assert_flash_error
+      assert_equal(old_count, spl.reload.observations.size)
+
+      login("mary")
+      put(:update,
+          params: params.merge(commit: :ADD.l, species_list: spl.title))
+      assert_response(:redirect)
+      assert_redirected_to(%r{/species_list/show_species_list})
+      assert_flash_success
+      assert_equal(new_count, spl.reload.observations.size)
+
+      put(:update,
+          params: params.merge(commit: :REMOVE.l, species_list: spl.id.to_s))
+      assert_response(:redirect)
+      assert_redirected_to(%r{/species_list/show_species_list})
+      assert_flash_success
+      assert_equal(0, spl.reload.observations.size)
     end
 
-    def test_manage_species_list_with_projects
-      proj = projects(:bolete_project)
-      spl1 = species_lists(:unknown_species_list)
-      spl2 = species_lists(:first_species_list)
-      spl3 = species_lists(:another_species_list)
-      spl2.user = dick
-      spl2.save
-      spl2.reload
-      obs1 = observations(:detailed_unknown_obs)
-      obs2 = observations(:coprinus_comatus_obs)
-      assert_obj_list_equal([dick], proj.user_group.users)
-      assert_obj_list_equal([proj], spl1.projects)
-      assert_obj_list_equal([], spl2.projects)
-      assert_obj_list_equal([], spl3.projects)
-      assert_true(spl1.observations.include?(obs1))
-      assert_false(spl1.observations.include?(obs2))
-      assert_obj_list_equal([], spl2.observations)
-      assert_obj_list_equal([], spl3.observations)
-      assert_users_equal(mary, spl1.user)
-      assert_users_equal(dick, spl2.user)
-      assert_users_equal(rolf, spl3.user)
-
-      login("dick")
-      get(:manage_species_lists, params: { id: obs1.id })
-      assert_select("a[href*='species_list=#{spl1.id}']",
-                    text: :REMOVE.t, count: 1)
-      assert_select("a[href*='species_list=#{spl2.id}']", text: :ADD.t,
-                                                          count: 1)
-      assert_select("a[href*='species_list=#{spl3.id}']", count: 0)
-
-      get(:manage_species_lists, params: { id: obs2.id })
-      assert_select("a[href*='species_list=#{spl1.id}']", text: :ADD.t,
-                                                          count: 1)
-      assert_select("a[href*='species_list=#{spl2.id}']", text: :ADD.t,
-                                                          count: 1)
-      assert_select("a[href*='species_list=#{spl3.id}']", count: 0)
-
-      post(:add_observation_to_species_list,
-           params: { observation: obs2.id,
-                     species_list: spl1.id })
-      assert_redirected_to(action: :manage_species_lists, id: obs2.id)
-      assert_true(spl1.reload.observations.include?(obs2))
-
-      post(:remove_observation_from_species_list,
-           params: { observation: obs2.id,
-                     species_list: spl1.id })
-      assert_redirected_to(action: :manage_species_lists, id: obs2.id)
-      assert_false(spl1.reload.observations.include?(obs2))
+    def test_post_add_remove_double_observations
+      spl = species_lists(:unknown_species_list)
+      old_obs_list =
+        SpeciesListObservation.select(:observation_id).
+        where(species_list: spl.id).
+        order(observation_id: :asc).
+        map(&:observation_id)
+      dup_obs = spl.observations.first
+      new_obs = (Observation.all - spl.observations).first
+      ids = [dup_obs.id, new_obs.id]
+      query = Query.lookup(:Observation, :in_set, ids: ids)
+      params = @controller.query_params(query).merge(
+        commit: :ADD.l,
+        species_list: spl.title
+      )
+      login(spl.user.login)
+      put(:update, params: params)
+      assert_response(:redirect)
+      assert_flash_success
+      new_obs_list =
+        SpeciesListObservation.select(:observation_id).
+        where(species_list: spl.id).
+        order(observation_id: :asc).
+        map(&:observation_id)
+      assert_equal(new_obs_list.length, old_obs_list.length + 1)
+      assert_equal((new_obs_list - old_obs_list).first, new_obs.id)
     end
   end
 end
